teste

+++ price-management-api/src/main/java/com/Cocacola/pricemanagementapi/Entities/Product.java
@@ -28,9 +28,9 @@ public class Product {
 
     private Integer quantity_ml;
 
-    private Long stock_quantity;
+    private Integer stock_quantity;
 
-    private Long stock_Max;
+    private Integer stock_Max;
 
     private double price;
